+++ src/test/java/dev/noctud/latte/lexer/LatteTopLexerAdapterTest.java
@@ -237,6 +237,276 @@ public class LatteTopLexerAdapterTest {
 
 	@Test
 	@SuppressWarnings("unchecked")
+	public void testSyntaxOff() {
+		Lexer lexer = new LatteTopLexerAdapter();
+
+		// Basic syntax off - macros inside become plain text
+		lexer.start("{syntax off}{$var}{/syntax}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax off}"),
+			Pair.create(T_TEXT, "{$var}"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+		});
+
+		// Syntax off with surrounding content
+		lexer.start("A{syntax off}B{$var}C{/syntax}D");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_TEXT, "A"),
+			Pair.create(T_MACRO_CLASSIC, "{syntax off}"),
+			Pair.create(T_TEXT, "B{$var}C"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+			Pair.create(T_TEXT, "D"),
+		});
+
+		// Comments also suppressed inside syntax off
+		lexer.start("{syntax off}{* comment *}{/syntax}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax off}"),
+			Pair.create(T_TEXT, "{* comment *}"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+		});
+
+		// Multiple macros suppressed
+		lexer.start("{syntax off}{if true}hello{/if}{/syntax}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax off}"),
+			Pair.create(T_TEXT, "{if true}hello{/if}"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+		});
+
+		// HTML tags inside syntax off are plain text
+		lexer.start("{syntax off}<div class=\"test\">{$var}</div>{/syntax}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax off}"),
+			Pair.create(T_TEXT, "<div class=\"test\">{$var}</div>"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+		});
+
+		// Unclosed syntax off - consumes to EOF
+		lexer.start("{syntax off}{$var}text");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax off}"),
+			Pair.create(T_TEXT, "{$var}text"),
+		});
+	}
+
+	@Test
+	@SuppressWarnings("unchecked")
+	public void testSyntaxDouble() {
+		Lexer lexer = new LatteTopLexerAdapter();
+
+		// Single braces are plain text, double braces are macros
+		// Closing with single-brace {/syntax} (Latte uses original delimiters for close)
+		lexer.start("{syntax double}{$var}{{$var}}{/syntax}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax double}"),
+			Pair.create(T_TEXT, "{$var}"),
+			Pair.create(T_MACRO_CLASSIC, "{{$var}}"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+		});
+
+		// Also works with double-brace close
+		lexer.start("{syntax double}{{$var}}{{/syntax}}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax double}"),
+			Pair.create(T_MACRO_CLASSIC, "{{$var}}"),
+			Pair.create(T_MACRO_CLASSIC, "{{/syntax}}"),
+		});
+
+		// Double-brace pair macros
+		lexer.start("{syntax double}{{if $cond}}hello{{/if}}{/syntax}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax double}"),
+			Pair.create(T_MACRO_CLASSIC, "{{if $cond}}"),
+			Pair.create(T_TEXT, "hello"),
+			Pair.create(T_MACRO_CLASSIC, "{{/if}}"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+		});
+
+		// Double-brace comments
+		lexer.start("{syntax double}{{* comment *}}text{/syntax}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax double}"),
+			Pair.create(T_MACRO_COMMENT, "{{* comment *}}"),
+			Pair.create(T_TEXT, "text"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+		});
+
+		// JavaScript-like content with single braces is preserved as text
+		lexer.start("{syntax double}var x = {a: 1, b: 2};{/syntax}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax double}"),
+			Pair.create(T_TEXT, "var x = {a: 1, b: 2};"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+		});
+
+		// HTML tags and n:attributes work inside syntax double
+		lexer.start("{syntax double}<span n:if=\"$show\">{{$var}}</span>{/syntax}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_MACRO_CLASSIC, "{syntax double}"),
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "span "),
+			Pair.create(T_HTML_TAG_NATTR_NAME, "n:if"),
+			Pair.create(T_HTML_TAG_ATTR_EQUAL_SIGN, "="),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_MACRO_CONTENT, "$show"),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			Pair.create(T_MACRO_CLASSIC, "{{$var}}"),
+			Pair.create(T_HTML_CLOSE_TAG_OPEN, "</"),
+			Pair.create(T_TEXT, "span"),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			Pair.create(T_MACRO_CLASSIC, "{/syntax}"),
+		});
+	}
+
+	@Test
+	@SuppressWarnings("unchecked")
+	public void testNSyntaxOff() {
+		Lexer lexer = new LatteTopLexerAdapter();
+
+		// Basic n:syntax="off" - macros inside become plain text
+		lexer.start("<div n:syntax=\"off\">{$var}</div>");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "div "),
+			Pair.create(T_HTML_TAG_NATTR_NAME, "n:syntax"),
+			Pair.create(T_HTML_TAG_ATTR_EQUAL_SIGN, "="),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_MACRO_CONTENT, "off"),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			Pair.create(T_TEXT, "{$var}"),
+			Pair.create(T_HTML_CLOSE_TAG_OPEN, "</"),
+			Pair.create(T_TEXT, "div"),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+		});
+
+		// n:syntax with other attributes
+		lexer.start("<div class=\"test\" n:syntax=\"off\">{if $x}y{/if}</div>");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "div class=\"test\" "),
+			Pair.create(T_HTML_TAG_NATTR_NAME, "n:syntax"),
+			Pair.create(T_HTML_TAG_ATTR_EQUAL_SIGN, "="),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_MACRO_CONTENT, "off"),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			Pair.create(T_TEXT, "{if $x}y{/if}"),
+			Pair.create(T_HTML_CLOSE_TAG_OPEN, "</"),
+			Pair.create(T_TEXT, "div"),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+		});
+
+		// n:syntax with nested same-name tags
+		lexer.start("<div n:syntax=\"off\"><div>inner</div></div>");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "div "),
+			Pair.create(T_HTML_TAG_NATTR_NAME, "n:syntax"),
+			Pair.create(T_HTML_TAG_ATTR_EQUAL_SIGN, "="),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_MACRO_CONTENT, "off"),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			Pair.create(T_TEXT, "<div>inner</div>"),
+			Pair.create(T_HTML_CLOSE_TAG_OPEN, "</"),
+			Pair.create(T_TEXT, "div"),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+		});
+
+		// Self-closing tag with n:syntax should not enter syntax off mode
+		lexer.start("<br n:syntax=\"off\" />{$var}");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "br "),
+			Pair.create(T_HTML_TAG_NATTR_NAME, "n:syntax"),
+			Pair.create(T_HTML_TAG_ATTR_EQUAL_SIGN, "="),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_MACRO_CONTENT, "off"),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_TEXT, " "),
+			Pair.create(T_HTML_OPEN_TAG_CLOSE, "/>"),
+			Pair.create(T_MACRO_CLASSIC, "{$var}"),
+		});
+
+		// n:syntax="off" on <script> tag - macros suppressed inside, next tag normal
+		lexer.start("<script n:syntax=\"off\">var x = {a: 1};</script><script>{$var}</script>");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "script "),
+			Pair.create(T_HTML_TAG_NATTR_NAME, "n:syntax"),
+			Pair.create(T_HTML_TAG_ATTR_EQUAL_SIGN, "="),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_MACRO_CONTENT, "off"),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			// </script close uses T_TEXT pattern (like SCRIPT_CDATA), not T_HTML_CLOSE_TAG_OPEN
+			Pair.create(T_TEXT, ">var x = {a: 1};</script"),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			// Second script tag works normally with macro parsing
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "script>"),
+			Pair.create(T_MACRO_CLASSIC, "{$var}"),
+			Pair.create(T_TEXT, "</script"),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+		});
+	}
+
+	@Test
+	@SuppressWarnings("unchecked")
+	public void testNSyntaxDouble() {
+		Lexer lexer = new LatteTopLexerAdapter();
+
+		// n:syntax="double" - single braces are text, double braces are macros
+		lexer.start("<div n:syntax=\"double\">{$test}{{$test}}</div>");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "div "),
+			Pair.create(T_HTML_TAG_NATTR_NAME, "n:syntax"),
+			Pair.create(T_HTML_TAG_ATTR_EQUAL_SIGN, "="),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_MACRO_CONTENT, "double"),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			Pair.create(T_TEXT, "{$test}"),
+			Pair.create(T_MACRO_CLASSIC, "{{$test}}"),
+			Pair.create(T_HTML_CLOSE_TAG_OPEN, "</"),
+			Pair.create(T_TEXT, "div"),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+		});
+
+		// n:attributes on tags inside n:syntax="double" should work
+		lexer.start("<div n:syntax=\"double\"><span n:if=\"$show\">{{$var}}</span></div>");
+		assertTokens(lexer, new Pair[] {
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "div "),
+			Pair.create(T_HTML_TAG_NATTR_NAME, "n:syntax"),
+			Pair.create(T_HTML_TAG_ATTR_EQUAL_SIGN, "="),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_MACRO_CONTENT, "double"),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			Pair.create(T_HTML_OPEN_TAG_OPEN, "<"),
+			Pair.create(T_TEXT, "span "),
+			Pair.create(T_HTML_TAG_NATTR_NAME, "n:if"),
+			Pair.create(T_HTML_TAG_ATTR_EQUAL_SIGN, "="),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_MACRO_CONTENT, "$show"),
+			Pair.create(T_HTML_TAG_ATTR_DQ, "\""),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			Pair.create(T_MACRO_CLASSIC, "{{$var}}"),
+			Pair.create(T_HTML_CLOSE_TAG_OPEN, "</"),
+			Pair.create(T_TEXT, "span"),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+			Pair.create(T_HTML_CLOSE_TAG_OPEN, "</"),
+			Pair.create(T_TEXT, "div"),
+			Pair.create(T_HTML_TAG_CLOSE, ">"),
+		});
+	}
+
+	@Test
+	@SuppressWarnings("unchecked")
 	public void testMacroInAttribute() {
 		Lexer lexer = new LatteTopLexerAdapter();